requirements.txt
asgiref==3.4.1
black==21.11b1
click==8.0.3
Django==3.2.9
django-cors-headers==3.10.0
djangorestframework==3.12.4
importlib-metadata==4.8.2
injector==0.18.4
mypy-extensions==0.4.3
mysqlclient>=2.0
pathspec==0.9.0
platformdirs==2.4.0
pytz==2021.3
regex==2021.11.10
sqlparse==0.4.2
tomli==1.2.2
typed-ast==1.5.0
typing-extensions==4.0.0
zipp==3.6.0
gunicorn==20.1.0
